move additionalStacks into custom and fix syntax and indentation errors
in order for additionalStacks plugin to work it must be in the custom section of the serverless.yml file. When I checked the AWS console a realised the appsync stack wasn't deployed I figured this out. Moving the section was the start of a series of syntax errors that needed to be corrected